circle: add env config `SKIP_CHART_APP_VERSION_UPDATE` to skip Chart `appVersion` field updates
For charts deploying more than one container image,
`SKIP_CHART_APP_VERSION_UPDATE=1` can be defined in the `circle.yml`
file of the secondary image repo so that the chart `appVersion` is
only updated with the primary image updates.

For example, the parse chart deploys the `bitnami/parse` image and
the `bitnami/parse-dashboard` image. For this chart the `appVersion`
should reflect the version of the Parse server deployed by the
`bitnami/parse` repo. To disable `appVersion` update by the
`bitnami/parse-dashboard` repo, we would have to add
`SKIP_CHART_APP_VERSION_UPDATE=1` to the CI environment (`circle.yml`)